morenorm

+++ philo_one/headers_philo_one.h
@@ -6,7 +6,7 @@
 /*   By: user42 <root@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/18 20:45:08 by user42            #+#    #+#             */
-/*   Updated: 2020/11/08 23:54:10 by user42           ###   ########.fr       */
+/*   Updated: 2020/11/12 16:19:31 by user42           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,11 +27,11 @@
 */
 
 typedef struct			s_timers {
-	int					eat;
+	unsigned long		eat;
 	unsigned long		eat_us;
-	int					sleep;
+	unsigned long		sleep;
 	unsigned long		sleep_us;
-	int					die;
+	unsigned long		die;
 	unsigned long		die_us;
 	unsigned long		start;
 }						t_timers;
